initial Commit
The code is open source then GitHub forks and pull requests are really appreciated.
This code was released because around 200 developers around the world ask for it in the last weeks. The scope is to share the conical dynamic deformation, finger tracking, shadow casting, page constrain logic and texture mapping. It means that it's not yet ready to be used in a commercial project.

Signed-off-by: Lorenzo Manfredi <lorenzo@mutado.com>